Forge: 64-bit game clock timestamps, no clock budget
The sim clock runs far faster than real time, and absolute timestamps kept as uint32 milliseconds (with the 30-day
infinityCooldownDelay added for event-started cooldowns) wrapped after ~19.7 game days, so ForgeUpdateLoop stopped
the sim there: a curriculum run is several times that. They are uint64 now: player and creature spell cooldowns,
creature school lockouts, gameobject cooldowns, Sanctuary, Strand of the Ancients demolishers, and the Eclipse and
turkey marker script timers. ForgeUpdateLoop no longer has a budget. Relative timers compared through getMSTimeDiff
stay uint32 and wrap safely; GlobalCooldownMgr::GetGlobalCooldown now computes the time left that way too, where
start + delay against now read no cooldown for up to a GCD at each wrap of the clock's low 32 bits.

+++ src/server/game/Battlegrounds/Zones/BattlegroundSA.cpp
@@ -1122,11 +1122,11 @@ void BattlegroundSA::UpdateDemolisherSpawns()
                     // Demolisher is not in list
                     if (DemoliserRespawnList.find(i) == DemoliserRespawnList.end())
                     {
-                        DemoliserRespawnList[i] = GameTime::GetGameTimeMS().count() + 30000;
+                        DemoliserRespawnList[i] = uint64(GameTime::GetGameTimeMS().count()) + 30000;
                     }
                     else
                     {
-                        if (DemoliserRespawnList[i] < GameTime::GetGameTimeMS().count())
+                        if (DemoliserRespawnList[i] < uint64(GameTime::GetGameTimeMS().count()))
                         {
                             Demolisher->Relocate(BG_SA_NpcSpawnlocs[i][0], BG_SA_NpcSpawnlocs[i][1],
                                                  BG_SA_NpcSpawnlocs[i][2], BG_SA_NpcSpawnlocs[i][3]);
