+++ src/server/game/Entities/Creature/Creature.cpp
@@ -2949,7 +2949,7 @@ void Creature::ProhibitSpellSchool(SpellSchoolMask idSchoolMask, uint32 unTimeMs
     {
         if (idSchoolMask & (1 << i))
         {
-            m_ProhibitSchoolTime[i] = GameTime::GetGameTimeMS().count() + unTimeMs;
+            m_ProhibitSchoolTime[i] = uint64(GameTime::GetGameTimeMS().count()) + unTimeMs;
         }
     }
 }
@@ -2960,7 +2960,7 @@ bool Creature::IsSpellProhibited(SpellSchoolMask idSchoolMask) const
     {
         if (idSchoolMask & (1 << i))
         {
-            if (m_ProhibitSchoolTime[i] >= GameTime::GetGameTimeMS().count())
+            if (m_ProhibitSchoolTime[i] >= uint64(GameTime::GetGameTimeMS().count()))
             {
                 return true;
             }
@@ -2982,7 +2982,7 @@ void Creature::_AddCreatureSpellCooldown(uint32 spell_id, uint16 categoryId, uin
 {
     CreatureSpellCooldown spellCooldown;
     spellCooldown.category = categoryId;
-    spellCooldown.end = GameTime::GetGameTimeMS().count() + end_time;
+    spellCooldown.end = uint64(GameTime::GetGameTimeMS().count()) + end_time;
     m_CreatureSpellCooldowns[spell_id] = std::move(spellCooldown);
 }
 
@@ -3038,13 +3038,13 @@ uint32 Creature::GetSpellCooldown(uint32 spell_id) const
     if (itr == m_CreatureSpellCooldowns.end())
         return 0;
 
-    return itr->second.end > GameTime::GetGameTimeMS().count() ? itr->second.end - GameTime::GetGameTimeMS().count() : 0;
+    return itr->second.end > uint64(GameTime::GetGameTimeMS().count()) ? itr->second.end - uint64(GameTime::GetGameTimeMS().count()) : 0;
 }
 
 bool Creature::HasSpellCooldown(uint32 spell_id) const
 {
     CreatureSpellCooldowns::const_iterator itr = m_CreatureSpellCooldowns.find(spell_id);
-    return (itr != m_CreatureSpellCooldowns.end() && itr->second.end > GameTime::GetGameTimeMS().count());
+    return (itr != m_CreatureSpellCooldowns.end() && itr->second.end > uint64(GameTime::GetGameTimeMS().count()));
 }
 
 bool Creature::HasSpell(uint32 spellID) const
