+++ src/server/game/Entities/GameObject/GameObject.cpp
@@ -489,9 +489,9 @@ void GameObject::Update(uint32 diff)
                             GameObjectTemplate const* goInfo = GetGOInfo();
                             // Bombs
                             if (goInfo->trap.type == 2)
-                                m_cooldownTime = GameTime::GetGameTimeMS().count() + 10 * IN_MILLISECONDS; // Hardcoded tooltip value
+                                m_cooldownTime = uint64(GameTime::GetGameTimeMS().count()) + 10 * IN_MILLISECONDS; // Hardcoded tooltip value
                             else if (GetOwner())
-                                m_cooldownTime = GameTime::GetGameTimeMS().count() + goInfo->trap.startDelay * IN_MILLISECONDS;
+                                m_cooldownTime = uint64(GameTime::GetGameTimeMS().count()) + goInfo->trap.startDelay * IN_MILLISECONDS;
 
                             m_lootState = GO_READY;
                             break;
@@ -523,7 +523,7 @@ void GameObject::Update(uint32 diff)
                         }
                     case GAMEOBJECT_TYPE_SUMMONING_RITUAL:
                         {
-                            if (GameTime::GetGameTimeMS().count() < m_cooldownTime)
+                            if (uint64(GameTime::GetGameTimeMS().count()) < m_cooldownTime)
                                 return;
                             GameObjectTemplate const* info = GetGOInfo();
 
@@ -678,7 +678,7 @@ void GameObject::Update(uint32 diff)
                     GameObjectTemplate const* goInfo = GetGOInfo();
                     if (goInfo->type == GAMEOBJECT_TYPE_TRAP)
                     {
-                        if (GameTime::GetGameTimeMS().count() < m_cooldownTime)
+                        if (uint64(GameTime::GetGameTimeMS().count()) < m_cooldownTime)
                             break;
 
                         // Type 2 - Bomb (will go away after casting it's spell)
@@ -748,11 +748,11 @@ void GameObject::Update(uint32 diff)
                 {
                     case GAMEOBJECT_TYPE_DOOR:
                     case GAMEOBJECT_TYPE_BUTTON:
-                        if (GetGOInfo()->GetAutoCloseTime() && GameTime::GetGameTimeMS().count() >= m_cooldownTime)
+                        if (GetGOInfo()->GetAutoCloseTime() && uint64(GameTime::GetGameTimeMS().count()) >= m_cooldownTime)
                             ResetDoorOrButton();
                         break;
                     case GAMEOBJECT_TYPE_GOOBER:
-                        if (GameTime::GetGameTimeMS().count() >= m_cooldownTime)
+                        if (uint64(GameTime::GetGameTimeMS().count()) >= m_cooldownTime)
                         {
                             RemoveGameObjectFlag(GO_FLAG_IN_USE);
 
@@ -798,7 +798,7 @@ void GameObject::Update(uint32 diff)
                             if (goInfo->trap.spellId)
                                 CastSpell(target, goInfo->trap.spellId);
 
-                            m_cooldownTime = GameTime::GetGameTimeMS().count() + (goInfo->trap.cooldown ? goInfo->trap.cooldown : uint32(4)) * IN_MILLISECONDS; // template or 4 seconds
+                            m_cooldownTime = uint64(GameTime::GetGameTimeMS().count()) + (goInfo->trap.cooldown ? goInfo->trap.cooldown : uint32(4)) * IN_MILLISECONDS; // template or 4 seconds
 
                             if (goInfo->trap.type == 1)
                                 SetLootState(GO_JUST_DEACTIVATED);
@@ -1421,7 +1421,7 @@ void GameObject::UseDoorOrButton(uint32 time_to_restore, bool alternative /* = f
     SwitchDoorOrButton(true, alternative);
     SetLootState(GO_ACTIVATED, user);
 
-    m_cooldownTime = GameTime::GetGameTimeMS().count() + time_to_restore;
+    m_cooldownTime = uint64(GameTime::GetGameTimeMS().count()) + time_to_restore;
 }
 
 void GameObject::SetGoArtKit(uint8 kit)
@@ -1483,10 +1483,10 @@ void GameObject::Use(Unit* user)
     // If cooldown data present in template
     if (uint32 cooldown = GetGOInfo()->GetCooldown())
     {
-        if (GameTime::GetGameTimeMS().count() < m_cooldownTime)
+        if (uint64(GameTime::GetGameTimeMS().count()) < m_cooldownTime)
             return;
 
-        m_cooldownTime = GameTime::GetGameTimeMS().count() + cooldown * IN_MILLISECONDS;
+        m_cooldownTime = uint64(GameTime::GetGameTimeMS().count()) + cooldown * IN_MILLISECONDS;
     }
 
     if (user->IsPlayer() && GetGoType() != GAMEOBJECT_TYPE_TRAP) // workaround for GO casting
@@ -1524,7 +1524,7 @@ void GameObject::Use(Unit* user)
                 if (goInfo->trap.spellId)
                     CastSpell(user, goInfo->trap.spellId);
 
-                m_cooldownTime = GameTime::GetGameTimeMS().count() + (goInfo->trap.cooldown ? goInfo->trap.cooldown :  uint32(4)) * IN_MILLISECONDS; // template or 4 seconds
+                m_cooldownTime = uint64(GameTime::GetGameTimeMS().count()) + (goInfo->trap.cooldown ? goInfo->trap.cooldown :  uint32(4)) * IN_MILLISECONDS; // template or 4 seconds
 
                 if (goInfo->trap.type == 1)         // Deactivate after trigger
                     SetLootState(GO_JUST_DEACTIVATED);
@@ -1691,7 +1691,7 @@ void GameObject::Use(Unit* user)
                 if (info->goober.customAnim)
                     SendCustomAnim(GetGoAnimProgress());
 
-                m_cooldownTime = GameTime::GetGameTimeMS().count() + info->GetAutoCloseTime();
+                m_cooldownTime = uint64(GameTime::GetGameTimeMS().count()) + info->GetAutoCloseTime();
 
                 // cast this spell later if provided
                 spellId = info->goober.spellId;
@@ -1868,7 +1868,7 @@ void GameObject::Use(Unit* user)
                     SetLootState(GO_NOT_READY);
 
                     // channel ready, maintain this
-                    m_cooldownTime = GameTime::GetGameTimeMS().count() + 5 * IN_MILLISECONDS;
+                    m_cooldownTime = uint64(GameTime::GetGameTimeMS().count()) + 5 * IN_MILLISECONDS;
                 }
 
                 return;
